add Acc benchmarks
I know Nikita Volkov as a Haskeller with a lot of smart logical solutions to
probems.

Acc is a tree-like structure which tree balances to the operation applied, so when a lot
of same operations applied to the structure - data tree is kept optimized for that
operation.

It is designed to be a builder-phase data type.

Therefore it lacks a lot of abilities.

Upstream thread with the author is:
https://github.com/nikita-volkov/acc/issues/1