add Snocing benchmarks
With datatypes that do efficient snocing - it is often easier to reason about &
build algorithms on, as data structure becomes FIFO.
As for example accumulator would add to the tail & then traversing the data
structure alalizing elements from old to new, which is easy to align with laziness.

DList & Acc also shine in snocing, that is also a main point to add snocing.