Revert "add Acc benchmarks"
This reverts commit f41659b1f2e3e3a0a6a1ed61577d5a6122d2beed.

Details are in the https://github.com/nikita-volkov/acc/issues/4

That thread lead me to request to revert my desigion.

In short: `Acc` hold the x2 faster then `{DList,Seq}` claim in the domain & way
that it asks to be used & benchmarks here so far are not granular or
integrative enough to represent that. & from current position there is a number
of steps needed to be able to present `Acc` properly - so it became it is
possible to remove the data type from testing while also it is impossible to
represent it properly - because people either don't have according skills or
time/effort priority to do the according work.

In the meantime of expanding the test suite step by stem - the data type would
be able to return under proper conditions, to become discoverable.